update delete buttun

+++ admin/delete_user.php
@@ -1,22 +1,21 @@
 <?php
+
 include "../db_connect.php";
 session_start();
-if(!isset($_SESSION["user"]) && !isset($_SESSION["admin"])){ 
-    header("Location: ../login/login.php"); 
+if (!isset($_SESSION["user"]) && !isset($_SESSION["admin"])) {
+    header("Location: ../login/login.php");
 }
 
-if(isset($_SESSION["user"])){ 
-    header("Location: ../home_user/home.php"); 
+if (isset($_SESSION["user"])) {
+    header("Location: ../home_user/home.php");
 }
 
-
-$sql="SELECT * FROM `users`";
-$result=mysqli_query($connect,$sql);
-$row=mysqli_fetch_assoc($result);
-$delete="DELETE FROM `users`";
-if(mysqli_query($connect,$delete)){
-header("Location: ../dashboard.php");
-}else{
-
+$id = $_GET["id"];
+$sql = "SELECT * FROM `users` WHERE user_id = $id";
+$result = mysqli_query($connect, $sql);
+$row = mysqli_fetch_assoc($result);
+$delete = "DELETE FROM `users` WHERE user_id = $id";
+if (mysqli_query($connect, $delete)) {
+    header("Location: dashboard_users.php");
+} else {
 }
-?>
